Convert floats straight to narrow integers and clamp with separate ifs
A float cast directly to u8/u16 zero-extends with movz, while going
through i32 first masks with and. A clamp written as an if after the
update is if-converted to a compare/and/or sequence, whereas a ternary
folds to minss/maxss.

+++ src/gameapi/edtools/edptlall.cpp
@@ -791,7 +791,7 @@ static void edptlcbApplyNumGhosts(eduimenu_s *, eduiitem_s *item, u32) {
     }
 
     debinftype *effect = debtab[debkeydata[instance_id].effect_index];
-    effect->trail_count = static_cast<u8>(static_cast<i32>(static_cast<edui_slider_s *>(item)->value));
+    effect->trail_count = static_cast<u8>(static_cast<edui_slider_s *>(item)->value);
     UpdateTotalPtls(effect);
 }
 static void cbPtlChangeIvalOffRan(eduimenu_s *, eduiitem_s *item, u32) {
